Update patch set 1
Patch Set 1:

> Hmm, if we are having echo cancelling issues? Then it should be fixed in the mixer paths. For example, my LeEco Le Max 2 had serious echo during phonecalls, but I did fix it in mixer paths and we still have that libqcomvoiceprocessing.so

The problems are in the *hardware AEC*. The audioeffects xml advertises "we have a hardware AEC filter in libqcomvoiceprocessing", but we deploy the androidbase libqcomvoiceprocessing, which is only a skeleton that does nothing. It's a skeleton for the vendor to fill in with code.

I don't think the mixer paths alone will work without a vendor-supplied libqcomvoiceprocessing implementation. Therefore, the solution I propose is, at least: don't advertise a working hardware AEC to the system. Then the system will choose software AEC as a solution making webrtc programs work.

A side path would be to pull in the vendor libqcomvoiceprocessing, but I failed to make that work too.

Patch-set: 1